Reorganize sync session tests
Changes:
- Moved sync session tests into their own directory
- Extracted some helper functions into their own set of utils files
- Extracted the progress notification tests into their own file
